Moved db interactions to own file
