+++ app/api/weave/route.ts
@@ -1,11 +1,10 @@
 
 //Route for actions on single weaves owned by a user
 import { auth } from '@clerk/nextjs';
-import { Db } from 'mongodb';
 import { NextResponse } from 'next/server';
 
-import { dbConnection } from '@/database/mongodb'
-import { Draft } from '@/database/types/documents'
+import { getDraft, saveDraft } from '../utils/handleWeave';
+
 export async function GET(
     req: Request,
 
@@ -17,16 +16,13 @@ export async function GET(
     }
 
     try {
-        const db = await dbConnection() as Db
-        let dbResponse = await db.collection('drafts').findOne({ userId: userId })
-        console.log(dbResponse)
-        if (!dbResponse) {
+        const draft = await getDraft(userId)
+
+        if (!draft) {
             return new NextResponse('You have no saved weaves', { status: 204 });
         }
-        console.log(dbResponse.weave)
-        let weaveObject = dbResponse.weave
+        return NextResponse.json({ weaveObject:draft }, { status: 200 });
 
-        return NextResponse.json({ weaveObject }, { status: 200 });
     } catch (error) {
         console.log('api/filehandler', error);
         return new NextResponse(
@@ -41,20 +37,17 @@ export async function POST(
     req: Request,
 
 ) {
-    
+
     const { userId } = auth();
     if (!userId) {
         return new NextResponse('Unauthorized', { status: 401 });
     }
 
     try {
-
-        const db = await dbConnection() as Db
         const body = await req.json();
-        const { weaveObject, user } = body.values
-        let newDocument: Draft = { userId: userId, name: 'One', weave: weaveObject, created: Date.now(), updated: Date.now(), public: false }
-        let dbResponse = await db.collection('drafts').insertOne(newDocument)
-        console.log(dbResponse)
+        const { weaveObject } = body.values
+        const draft = await saveDraft(userId, weaveObject)
+
 
         return NextResponse.json('Successfully posted WEAVE at route api/weaves/', { status: 201 });
     } catch (error) {
